refactor: split SessionBuilder into sub-components
