+++ src/components/professor/SessionBuilder/index.ts
@@ -1,0 +1,1 @@
+export { SessionBuilder } from "./SessionBuilder";
